Create worldometer.py and remove redundant modules
