--- coronainfo/models/__init__.py
@@ -1,1 +1,0 @@
-from .model_corona import CoronaData, CoronaHeaders
