+++ requirements.txt
@@ -1,2 +1,3 @@
 beautifulsoup4==4.11.1
 PyGObject==3.42.2
+requests==2.28.1
